ADD custom getter and setter for JobList jobs
Setter takes the String that is inputted by the user and creates an array
of Job objects from them. The new getter gets all the Jobs and
concatenates their job_id's together to output. There is no
dependancy/sorting logic at this stage; Jobs are outputted in the order
they were inputted