ADD logic to check for simple circular dependancy
This commit satisfies the simple circular dependancy example
a => b
b => c
c => a
The circular_dependancy?() method iterates over each Job object we created
that has a dependancy looking for a circular root in a Hash. This is done
by passing the job_id and the dependancy_id in turn of each Job to a
recursive_find() helper method. The method checks the value associated to
the job_id's and dependancy_id's key in the Hash; returning the id arg if
the value is nil or recursivley calling itself with keys value. This works
its way through the hash eventually finding a circular depndancy if the
two results of recursive_find are the same (i.e. the recursive_find result
for dependancy_id ends up being the same as the job_id of the Job)